use key type and key user as part of the internal keyId
splitting the key handle up into 3 parts internally, while presenting a one byte
key id to the user, allows the same external or user facing key to be identical
for each user value and between SHE and crypto keys. algorithm types were added
as flags instead of being a part of the key handle and the user param was added
to the client so a user can split operations between 16 clients. may not fully
implement multiple client handling but seems good for keys