Implement Key Wrap Demo
Implement key wrap demo that can provision the flash with a wrapped RSA
and AES GCM key as well as pull those keys out from flash and attempt to
decrypt a blob of data.

This also refactors the key wrap demo code to use the test suite code
directly so we don't have repeated code.