Implement SHA oneshot handling between client and server
This changes the way that the server and client process and handle SHA
requests to implement the ability to do oneshot SHA requests. Originally,
the server-client interface could only send block sized requests back
and forth relying on calls to update and finalize. With this PR it
changes the request layout to allow for a variable sized input data for
the server to take in and perform a SHA on.

This PR is necessary to properly support oneshot SHA Crypto HW acceleration

This PR also simplifies some of the SHA client code, removing some
redundancies.